+++ podpointclient/__init__.py
@@ -2,8 +2,9 @@
 
 from .client import PodPointClient
 from .domain import (
-    AccountCapability, BasicChargingMode, BoostState, CapabilitySupport, ChargeSession,
-    ChargerCapability, ChargerDomain, ChargerIdentityError, ChargerRef,
+    AccountCapability, BasicChargingMode, BoostState, CapabilitySupport,
+    ChargeSession, ChargeSessionSource, ChargerCapability, ChargerDomain,
+    ChargerIdentityError, ChargerRef,
     ChargerSource, ChargerState, NormalizedStateValue, StateValue,
     boost_state_from_home, boost_state_from_legacy, charger_ref_from_charger,
     charger_ref_from_pod, charge_session_from_home, charge_session_from_legacy,
@@ -13,7 +14,7 @@ from .errors import UnsupportedCapabilityError
 
 __all__ = [
     "AccountCapability", "BasicChargingMode", "BoostState", "CapabilitySupport",
-    "ChargeSession",
+    "ChargeSession", "ChargeSessionSource",
     "ChargerCapability", "ChargerDomain", "ChargerIdentityError", "ChargerRef",
     "ChargerSource", "ChargerState", "NormalizedStateValue", "PodPointClient",
     "StateValue", "UnsupportedCapabilityError", "boost_state_from_home",